bnplasso: Fit linear regression models with the Nonparametric Bayesian Lasso

Fit linear regression models with the Nonparametric Bayesian Lasso, a novel and adaptive shrinkage prior for Bayesian regression and variable selection. Posterior inference is conducted via Markov chain Monte Carlo (MCMC) as described in Marin et al. (2025+) <https://doi.org/10.1080/10618600.2025.2572327>.

References

S. Marin, B. Long,and A. H. Westveld (2025+), Adaptive Shrinkage with a Nonparametric Bayesian Lasso. Journal of Computational and Graphical Statistics. doi:10.1080/10618600.2025.2572327

See also

Author

Maintainer: Santiago Marin santiago.marinardila@anu.edu.au (ORCID)

Other contributors:

Bronwyn Loong bronwyn.loong@anu.edu.au (ORCID) [contributor]

Anton H. Westveld anton.westveld@anu.edu.au (ORCID) [contributor]